WS-2595 Permit successful gradlew build on Red Hat 7 and Oracle java 1.8
Bumped gradlew to 2.3, which bundles groovy 2.3.9. This was the earliest
version to support sdk15on (supporting Java SDK 1.8) and v 1.61 which
fixes the jar signing to not use the now-obsolete SHA-1 algorithm.

Bumped commons to also use this higher version of groovy (2.3.9) in
order to avoid groovy test compilation failures.

Bumped spock up to use a version supporting groovy 2.3.

Removed references to GroovyTestCase which is no longer present in
junit4 and higher.